+++ src/main/gen/com/github/nthykier/debpkg/deb822/deplang/psi/DepLangPackageName.java
@@ -4,10 +4,13 @@ package com.github.nthykier.debpkg.deb822.deplang.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiReference;
 
 public interface DepLangPackageName extends PsiElement {
 
   @Nullable
   DepLangSubstvar getSubstvar();
+
+  @Nullable PsiReference getReference();
 
 }